memory

+++ expconfigs/memory_multi_atlas_revunet.py
@@ -25,7 +25,7 @@ class ExpConfig():
         self.channels = [64, 128, 256, 512]
         self.channels = [int(x/32) for x in self.channels]
         self.net = RevUnet3D(1, self.channels, 12, interpolation = (512,512,198))
-        self.net.half()
+        
         # self.net = RevUnet3D(1, self.channels, 12, interpolation = (256,256,99))
 
         # Data
